Refactor: split out device management API into a seperate header
