+++ CHANGES
@@ -9,6 +9,8 @@ differs as follows:
 
 Changes for 1.0.0:
 
+- The API for managing devices and profiles is split out into a seperate header
+  edgex/device-mgmt.h
 - The device name is passed to the Get and Set handlers. The commandrequest
   structure is simplified.
 - Example implementation includes setting service name and registry location
