+++ include/edgex/devsdk.h
@@ -14,33 +14,9 @@
  * @brief This file defines the functions and callbacks relating to the SDK.
  */
 
-#include "edgex/edgex.h"
+#include "edgex/edgex-base.h"
 #include "edgex/error.h"
 #include "edgex/edgex_logging.h"
-#include "edgex/registry.h"
-
-typedef struct edgex_blob
-{
-  size_t size;
-  uint8_t *bytes;
-} edgex_blob;
-
-typedef union edgex_device_resultvalue
-{
-  bool bool_result;
-  char *string_result;
-  uint8_t ui8_result;
-  uint16_t ui16_result;
-  uint32_t ui32_result;
-  uint64_t ui64_result;
-  int8_t i8_result;
-  int16_t i16_result;
-  int32_t i32_result;
-  int64_t i64_result;
-  float f32_result;
-  double f64_result;
-  edgex_blob binary_result;
-} edgex_device_resultvalue;
 
 /**
  * @brief Structure containing information about a device resource which is
@@ -94,8 +70,8 @@ typedef bool (*edgex_device_device_initialize)
 );
 
 /**
- * @brief Request to dynamically discover devices. If the implementation is
- *        capable of doing so, it should detect devices and register them
+ * @brief Optional callback for dynamic discovery of devices. The
+ *        implementation should detect devices and register them
  *        using the edgex_device_add_device API call.
  * @param impl The context data passed in when the service was created.
  */
@@ -156,7 +132,7 @@ typedef bool (*edgex_device_handle_put)
 typedef bool (*edgex_device_disconnect_device)
 (
   void *impl,
-  edgex_addressable *device
+  edgex_protocols *device
 );
 
 /**
@@ -172,7 +148,7 @@ typedef void (*edgex_device_stop) (void *impl, bool force);
 typedef struct edgex_device_callbacks
 {
   edgex_device_device_initialize init;
-  edgex_device_discover discover;
+  edgex_device_discover discover;            /* NULL for no discovery */
   edgex_device_handle_get gethandler;
   edgex_device_handle_put puthandler;
   edgex_device_disconnect_device disconnect;
@@ -264,135 +240,5 @@ void edgex_device_service_stop
   bool force,
   edgex_error *err
 );
-
-/**
- * @brief Add a device to the EdgeX system. This will generally be called in
- *        response to a request for device discovery.
- * @param svc The device service.
- * @param name The name of the new device.
- * @param description Optional description of the new device.
- * @param labels Optional labels for the new device.
- * @param profile_name Name of the device profile to be used with this device.
- * @param address Addressable for this device. The addressable will be created
- *        in metadata. The address' name and origin timestamp will be generated
- *        if not set.
- * @param err Nonzero reason codes will be set here in the event of errors.
- * @returns The id of the newly created or existing device, or NULL if an error
- *          occurred.
- */
-
-char * edgex_device_add_device
-(
-  edgex_device_service *svc,
-  const char *name,
-  const char *description,
-  const edgex_strings *labels,
-  const char *profile_name,
-  edgex_protocols *protocols,
-  edgex_error *err
-);
-
-/**
- * @brief Remove a device from EdgeX. The device will be deleted from the
- *        device service and from core-metadata.
- * @param svc The device service.
- * @param id The id of the device to be removed.
- * @param err Nonzero reason codes will be set here in the event of errors.
- */
-
-void edgex_device_remove_device
-  (edgex_device_service *svc, const char *id, edgex_error *err);
-
-/**
- * @brief Remove a device from EdgeX. The device will be deleted from the
- *        device service and from core-metadata.
- * @param svc The device service.
- * @param name The name of the device to be removed.
- * @param err Nonzero reason codes will be set here in the event of errors.
- */
-
-void edgex_device_remove_device_byname
-  (edgex_device_service *svc, const char *name, edgex_error *err);
-
-/**
- * @brief Update a device's details.
- * @param svc The device service.
- * @param id The id of the device to update. If this is unset, the device will
- *           be located by name.
- * @param name If id is unset, this parameter must be set to the name of the
- *             devce to be updated. Otherwise, it is optional and if set,
- *             specifies a new name for the device.
- * @param description If set, a new description for the device.
- * @param labels If set, a new set of labels for the device.
- * @param profilename If set, a new device profile for the device.
- * @param err Nonzero reason codes will be set here in the event of errors.
- */
-
-void edgex_device_update_device
-(
-  edgex_device_service *svc,
-  const char *id,
-  const char *name,
-  const char *description,
-  const edgex_strings *labels,
-  const char *profilename,
-  edgex_error *err
-);
-
-/**
- * @brief Obtain a list of devices known to the system.
- * @param svc The device service.
- */
-
-edgex_device * edgex_device_devices (edgex_device_service *svc);
-
-/**
- * @brief Retrieve device information.
- * @param svc The device service.
- * @param id The device id.
- * @returns The requested device metadata or null if the device was not found.
- */
-
-edgex_device * edgex_device_get_device
-  (edgex_device_service *svc, const char *id);
-
-/**
- * @brief Retrieve device information.
- * @param svc The device service.
- * @param name The device name.
- * @returns The requested device metadata or null if the device was not found.
- */
-
-edgex_device * edgex_device_get_device_byname
-  (edgex_device_service *svc, const char *name);
-
-/**
- * @brief Free a device structure or list of device structures.
- * @param The device or the first device in the list.
- */
-
-void edgex_device_free_device (edgex_device *e);
-
-/**
- * @brief Retrieve the device profiles currently known in the SDK.
- * @param svc The device service.
- * @param profiles Is set to an array of device profiles. This should be
- *        freed using edgex_deviceprofile_free_array() after use.
- * @returns the size of the returned array.
- */
-
-uint32_t edgex_device_service_getprofiles
-(
-  edgex_device_service *svc,
-  edgex_deviceprofile **profiles
-);
-
-/**
- * @brief Free an array of device profiles.
- * @param e The array.
- * @param count The number of elements in the array.
- */
-
-void edgex_deviceprofile_free_array (edgex_deviceprofile *e, unsigned count);
 
 #endif
